Renamed state.js -> model.js
